Implementation/omar.guerrero/listview1 (#5)
* Second Activity feature

The second activity was added for implementation on the second buttton "start second activity".

* Base Adapter implementation

Implementation MyBaseAdapter and MyAdapter adapted to the new implementation.

* Glide implementation

The glide optional  implementation was added to the project on the SecondActivity.

* ListView implementation

How to use a ListView in an ListActivity. You use the predefined ArrayAdapter class and an existing Android layout for the rows.