Implementation/omar.guerrero/multi pane v2 (#9)
* Landscape Portrait implementation

The RssfeedActivity should show two fragments (two panes) if started in landscape mode. If it is started in portrait mode only one fragment should be shown. In this exercise you adjust your application to support this.

* Multi-Pane v2 requested improvements

the arguments setter were moved to DetailsFragment and pass only the arguments in the constructor.

* Revert "Multi-Pane v2 requested improvements"

This reverts commit d893a74ccf44f74f4dbe276b1ff79f1853e0bf05.

* Update improvements

Create a newInstance function to apply arguments.